Desktop/projects/Fitapp-fixierun-full/FitApp-Consolidated/backend/activity-tracking/google_fit/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
click>=8.1.7
python-dotenv>=1.0.0
tenacity>=8.2.3
python-dateutil>=2.8.2

